requirements.txt
joblib==1.0.1
numpy==1.19.5
pandas==1.1.5
pkg_resources==0.0.0
python-dateutil==2.8.2
pytz==2021.1
scikit-learn==0.24.2
scikit-surprise==1.1.1
scipy==1.5.4
six==1.16.0
sklearn==0.0
threadpoolctl==2.2.0
